+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 > Todos los cambios significativos en la librería serán registrados en éste documento.
 
 
-### Nuevo 
-
+### Nuevo
+* Eliminado `process()`
 * Métodos `__construct`, `get_ip`, `payment`.
 * `DOCUMENTACION.md`